Refactor recipe routes for improved structure and functionality
- Implemented personalization routes with reduced nesting depth and enhanced request enrichment.
- Developed shopping list routes for multi-recipe optimization and ingredient consolidation.
- Established translation routes for single and batch recipe translations, including support for multiple languages.
- Added comprehensive integration tests for all recipe routes, ensuring functionality, error handling, and route registration.
- Improved code quality by extracting helper functions and ensuring proper documentation across route modules.
- Updated database schema to reflect changes in recipe handling and user preferences.
